\name{comcat}
\alias{comcat}
\title{
  Echo data in a human readable form.
}
\description{
  
  Format data in the most appropriate human readable form.
  
}
\usage{
comcat(x, ...)
}
\arguments{
  
  \item{x}{object.}

  \item{...}{additional arguments passed on to format.}

}
\references{Package home page: \url{https://togaware.com/projects/rattle/}}
\author{\email{Graham.Williams@togaware.com}}

\examples{
  comcat(dim(iris))
}
