\name{find_fewest_levels}
\alias{find_fewest_levels}
\title{
  Identify the categoric columns with the fewest values.
}
\description{

  Return the name of the categoric columns with the fewest values.

}
\usage{
find_fewest_levels(df)
}
\arguments{

  \item{df}{A data frame.}
}
\details{

  A function to identify the categorical column with the fewest unique
  values This is used by Rattle V6 to identify potential target
  variables.

}
\keyword{metadata}
