\name{ggVarImp}
\alias{ggVarImp}
\title{
  Model.
}
\description{
  
  Model.
  
}
\usage{
ggVarImp(model, ...)
}
\arguments{
  
  \item{model}{object.}

  \item{...}{arguments passed on.}

}
\references{Package home page: \url{https://togaware.com/projects/rattle/}}
\author{\email{Graham.Williams@togaware.com}}

\examples{
\dontrun{ggVarImp(model)}
}
