\name{rattle_gui}
\alias{ISO_8601_to_POSIX_datetime_format}
\alias{ada.formula}
\alias{cat_toggled}
\alias{close_rattle}
\alias{con_toggled}
\alias{display_click_execute_message}
\alias{display_click_odbc_execute_message}
\alias{gtkmain_handler}
\alias{gtkmainquit_handler}
\alias{imp_edited}
\alias{imp_toggled}
\alias{load_rdata_set_combo}
\alias{on_glm_gaussian_radiobutton_toggled}
\alias{on_glm_linear_radiobutton_toggled}
\alias{on_glm_logistic_radiobutton_toggled}
\alias{on_glm_multinomial_radiobutton_toggled}
\alias{on_about_menu_activate}
\alias{on_ada_continue_button_clicked}
\alias{on_ada_defaults_button_clicked}
\alias{on_ada_draw_button_clicked}
\alias{on_ada_errors_button_clicked}
\alias{on_ada_importance_button_clicked}
\alias{on_ada_list_button_clicked}
\alias{on_ada_stumps_button_clicked}
\alias{on_ada_stumps_checkbutton_toggled}
\alias{on_arff_filechooserbutton_update_preview}
\alias{on_arff_radiobutton_toggled}
\alias{on_associate_plot_frequency_button_clicked}
\alias{on_associate_rules_button_clicked}
\alias{on_boost_radiobutton_toggled}
\alias{on_categorical_clear_button_clicked}
\alias{on_cleanup_radiobutton_toggled}
\alias{on_confusion_radiobutton_toggled}
\alias{on_continuous_clear_button_clicked}
\alias{on_copy1_activate}
\alias{on_correlation_radiobutton_toggled}
\alias{on_csv_filechooserbutton_update_preview}
\alias{on_csv_radiobutton_toggled}
\alias{on_ctivate}
\alias{on_cut1_activate}
\alias{on_data_arff_radiobutton_toggled}
\alias{on_data_csv_radiobutton_toggled}
\alias{on_data_edit_button_clicked}
\alias{on_data_entry_radiobutton_toggled}
\alias{on_data_filechooserbutton_file_set}
\alias{on_data_library_radiobutton_toggled}
\alias{on_data_odbc_radiobutton_toggled}
\alias{on_data_rdata_radiobutton_toggled}
\alias{on_data_rdataset_radiobutton_toggled}
\alias{on_data_view_button_clicked}
\alias{on_delete_menu_activate}
\alias{on_dtree_radiobutton_toggled}
\alias{on_e1071_radiobutton_toggled}
\alias{on_editdata_button_clicked}
\alias{on_evaluate_csv_radiobutton_toggled}
\alias{on_evaluate_distributions_button_clicked}
\alias{on_evaluate_radiobutton_group_changed}
\alias{on_evaluate_rdataset_radiobutton_toggled}
\alias{on_execute_button_clicked}
\alias{on_explot_radiobutton_toggled}
\alias{on_export_button_clicked}
\alias{on_export_pmml_activate}
\alias{on_gbm_importance_button_clicked}
\alias{on_ggobi_radiobutton_toggled}
\alias{on_hclust_data_plot_button_clicked}
\alias{on_hclust_dendrogram_button_clicked}
\alias{on_hclust_discriminant_plot_button_clicked}
\alias{on_hclust_plot_button_clicked}
\alias{on_hclust_radiobutton_toggled}
\alias{on_hclust_stats_button_clicked}
\alias{on_help_ada_activate}
\alias{on_help_arff_activate}
\alias{on_help_confusion_table_activate}
\alias{on_help_correlation_activate}
\alias{on_help_csv_activate}
\alias{on_help_distributions_activate}
\alias{on_help_general_activate}
\alias{on_help_ggobi_activate}
\alias{on_help_glm_activate}
\alias{on_help_hierarchical_correlation_activate}
\alias{on_help_impute_activate}
\alias{on_help_kmeans_activate}
\alias{on_help_log_activate}
\alias{on_help_nomenclature_data_activate}
\alias{on_help_normalise_activate}
\alias{on_help_odbc_activate}
\alias{on_help_principal_components_activate}
\alias{on_help_randomForest_activate}
\alias{on_help_rdata_file_activate}
\alias{on_help_rdataset_activate}
\alias{on_help_roles_activate}
\alias{on_help_rpart_activate}
\alias{on_help_sample_activate}
\alias{on_help_sensitivity_activate}
\alias{on_help_summary_activate}
\alias{on_help_support_vector_machine_activate}
\alias{on_help_weight_calculator_activate}
\alias{on_hiercor_radiobutton_toggled}
\alias{on_impute_constant_radiobutton_toggled}
\alias{on_impute_radiobutton_toggled}
\alias{on_kernlab_radiobutton_toggled}
\alias{on_kmeans_data_plot_button_clicked}
\alias{on_kmeans_discriminant_plot_button_clicked}
\alias{on_kmeans_hclust_centers_checkbutton_toggled}
\alias{on_kmeans_iterate_checkbutton_toggled}
\alias{on_kmeans_plot_button_clicked}
\alias{on_kmeans_radiobutton_toggled}
\alias{on_kmeans_seed_button_clicked}
\alias{on_kmeans_stats_button_clicked}
\alias{on_libdata_combobox_changed}
\alias{on_libdata_radiobutton_toggled}
\alias{on_lift_radiobutton_toggled}
\alias{on_new_activate}
\alias{on_new_button_clicked}
\alias{on_nnet_radiobutton_toggled}
\alias{on_normalise_radiobutton_toggled}
\alias{on_notebook_switch_page}
\alias{on_odbc_radiobutton_toggled}
\alias{on_open_activate}
\alias{on_open_button_clicked}
\alias{on_paste1_activate}
\alias{on_plot_close_button_clicked}
\alias{on_plot_copy_button_clicked}
\alias{on_plot_print_button_clicked}
\alias{on_plot_save_button_clicked}
\alias{on_prcomp_radiobutton_toggled}
\alias{on_precision_radiobutton_toggled}
\alias{on_priors_entry_changed}
\alias{on_rattle_menu_activate}
\alias{on_rdata_filechooserbutton_update_preview}
\alias{on_rdata_radiobutton_toggled}
\alias{on_rdataset_radiobutton_toggled}
\alias{on_regression_paradigm_radiobutton_toggled}
\alias{on_regression_radiobutton_toggled}
\alias{on_remap_radiobutton_toggled}
\alias{on_rf_errors_button_clicked}
\alias{on_rf_importance_button_clicked}
\alias{on_rf_print_tree_button_clicked}
\alias{on_rf_radiobutton_toggled}
\alias{on_risk_comboboxentry_changed}
\alias{on_risk_radiobutton_toggled}
\alias{on_roc_radiobutton_toggled}
\alias{on_rpart_best_radiobutton_toggled}
\alias{on_rpart_build_radiobutton_toggled}
\alias{on_rpart_loss_comboboxentry_set_focus_child}
\alias{on_rpart_plot_button_clicked}
\alias{on_rpart_rules_button_clicked}
\alias{on_rpart_tune_radiobutton_toggled}
\alias{on_sample_checkbutton_toggled}
\alias{on_sample_count_spinbutton_changed}
\alias{on_sample_percentage_spinbutton_changed}
\alias{on_sample_radiobutton_toggled}
\alias{on_sample_seed_button_clicked}
\alias{on_save_as_activate}
\alias{on_save_button_clicked}
\alias{on_save_menu_activate}
\alias{on_score_radiobutton_toggled}
\alias{on_sensitivity_radiobutton_toggled}
\alias{on_summary_find_button_clicked}
\alias{on_summary_next_button_clicked}
\alias{on_summary_radiobutton_toggled}
\alias{on_svm_kernel_comboboxentry_changed}
\alias{on_svm_radiobutton_toggled}
\alias{on_tools_associate_activate}
\alias{on_tools_cluster_activate}
\alias{on_tools_data_activate}
\alias{on_tools_evaluate_activate}
\alias{on_tools_explore_activate}
\alias{on_tools_log_activate}
\alias{on_tools_model_activate}
\alias{on_tools_sample_activate}
\alias{on_tools_transform_activate}
\alias{on_tools_variables_activate}
\alias{on_tooltips_activate}
\alias{on_twoclass_radiobutton_toggled}
\alias{on_unsupervised_radiobutton_toggled}
\alias{on_variables_toggle_ignore_button_clicked}
\alias{on_variables_toggle_input_button_clicked}
\alias{on_viewdata_button_clicked}
\alias{on_viewdata_find_button_clicked}
\alias{on_viewdata_next_button_clicked}
\alias{open_odbc_set_combo}
\alias{quit_rattle}
\alias{update_comboboxentry_with_dataframes}
\alias{update_libdata_combobox_entry}
\title{Interal Rattle user interface callbacks.}
\description{These are exported from the package so that the GUI can
  access the callbacks. They should otherwise be ignored.}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
\keyword{internal}
