\name{listAdaVarsUsed}
\alias{listAdaVarsUsed}
\title{
  List the variables used by an adaboost model
}
\description{
  
  Returns a list of the variables used and their frequencies.
  
}
\usage{
listAdaVarsUsed(model)

}
\arguments{
  
  \item{model}{an rpart object.}

}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
