\name{modalvalue}
\alias{modalvalue}
\title{
  Calculate the mode of a vector, array or list.
}
\description{
  
   The mode is the most common or modal value of a list.
  
}
\usage{
modalvalue(x, na.rm=FALSE)
}
\arguments{
  
  \item{x}{A vector, array or list.}

  \item{na.rm}{Whether to remove missing values.}
}
\details{

  This function calculates the mode of a vector, array or list (lists
   are flattened). This code originated from an anonymous post on the R
   Wiki.

}
\keyword{hplot}
