\name{rattle.info}
\alias{rattle.info}
\title{  Print version information for Rattle and for R.}
\description{

  Display system information, including versions of Rattle and R,
  operating system, etc. Useful for reporting bugs.

}
\usage{
rattle.info()
}
\details{

  This is a support function to list useful information to provide the
  developers with information about the system environment when running
  Rattle. It is intended to provide the information that is useful in
  reporting bugs.

}

\references{Package home page: \url{http://rattle.togaware.com}}

\author{\email{Graham.Williams@togaware.com}}

\seealso{

  \code{\link{rattle}}.

}

\keyword{environment}
