\name{theme_rattle}
\alias{theme_rattle}
\title{
  A default GGPlot2 theme for Rattle V6
}
\description{

  This is a GGPlot2 theme for Rattle V6

}
\usage{
theme_rattle(base_size, base_family)
}
\arguments{

  \item{base_size}{Font size to use. Default is 11.}

  \item{base_family}{Font family to use.}
}
\details{

  This is a default for the Rattle V6 GGPlot2 theme for plots.

}
\keyword{theme}
