\name{whichNumerics}
\alias{whichNumerics}
\title{
  Returns a list of the names of the numeric variables in a data frame.
}
\description{
  
  A rattle support function.
  
}
\usage{
whichNumerics(data)

}
\arguments{
  
  \item{data}{a data frame.}

}
\references{Package home page: \url{https://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
