\name{binning}
\alias{binning}
\title{
  Perform binning
}
\description{
  
  Perform binning.
  
}
\usage{
binning(x, bins=4, method=c("quantile", "kmeans"),
                     labels=NULL, ordered=TRUE)
}
\arguments{
  
  \item{x}{data.}

  \item{bins}{number of bins.}

  \item{method}{mechanism.}

  \item{labels}{labels.}

  \item{ordered}{oredered.}
  
}
\details{

  Perform binning
  
}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
