\name{ggVarImp}
\alias{ggVarImp}
\title{
  Model.
}
\description{
  
  Model.
  
}
\usage{
ggVarImp(model, ...)
}
\arguments{
  
  \item{model}{object.}

  \item{...}{arguments passed on.}

}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}

\examples{
\dontrun{ggVarImp(model)}
}
