\name{rattle_gui}
\alias{cat_toggled}
\alias{con_toggled}
\alias{display_click_execute_message}
\alias{load_rdata_set_combo}
\alias{on_about_menu_activate}
\alias{on_boost_radiobutton_toggled}
\alias{on_categorical_clear_button_clicked}
\alias{on_confusion_radiobutton_toggled}
\alias{on_continuous_clear_button_clicked}
\alias{on_copy1_activate}
\alias{on_correlation_radiobutton_toggled}
\alias{on_csv_filechooserbutton_update_preview}
\alias{on_csv_radiobutton_toggled}
\alias{on_ctivate}
\alias{on_cut1_activate}
\alias{on_delete_menu_activate}
\alias{on_dtree_radiobutton_toggled}
\alias{on_e1071_radiobutton_toggled}
\alias{on_evaluate_csv_radiobutton_toggled}
\alias{on_evaluate_distributions_button_clicked}
\alias{on_evaluate_radiobutton_group_changed}
\alias{on_evaluate_rdataset_radiobutton_toggled}
\alias{on_execute_button_clicked}
\alias{on_explot_radiobutton_toggled}
\alias{on_export_activate}
\alias{on_export_pmml_activate}
\alias{on_ggobi_radiobutton_toggled}
\alias{on_hclust_radiobutton_toggled}
\alias{on_help_confusion_table_activate}
\alias{on_help_correlation_activate}
\alias{on_help_csv_activate}
\alias{on_help_gbm_activate}
\alias{on_help_general_activate}
\alias{on_help_glm_activate}
\alias{on_help_hierarchical_correlation_activate}
\alias{on_help_kmeans_activate}
\alias{on_help_nomenclature_data_activate}
\alias{on_help_odbc_activate}
\alias{on_help_principal_components_activate}
\alias{on_help_randomForest_activate}
\alias{on_help_rdata_file_activate}
\alias{on_help_rdataset_activate}
\alias{on_help_roles_activate}
\alias{on_help_rpart_activate}
\alias{on_help_sensitivity_activate}
\alias{on_help_summary_activate}
\alias{on_help_support_vector_machine_activate}
\alias{on_help_weight_calculator_activate}
\alias{on_hiercor_radiobutton_toggled}
\alias{on_kernlab_radiobutton_toggled}
\alias{on_kmeans_radiobutton_toggled}
\alias{on_lift_radiobutton_toggled}
\alias{on_new1_activate}
\alias{on_new_button_clicked}
\alias{on_notebook_switch_page}
\alias{on_odbc_radiobutton_toggled}
\alias{on_open_activate}
\alias{on_open_button_clicked}
\alias{on_paste1_activate}
\alias{on_prcomp_radiobutton_toggled}
\alias{on_precision_radiobutton_toggled}
\alias{on_priors_entry_changed}
\alias{on_rdata_filechooserbutton_update_preview}
\alias{on_rdata_radiobutton_toggled}
\alias{on_rdataset_radiobutton_toggled}
\alias{on_regression_radiobutton_toggled}
\alias{on_rf_radiobutton_toggled}
\alias{on_risk_comboboxentry_changed}
\alias{on_risk_radiobutton_toggled}
\alias{on_roc_radiobutton_toggled}
\alias{on_rpart_loss_comboboxentry_set_focus_child}
\alias{on_sample_checkbutton_toggled}
\alias{on_sample_count_spinbutton_changed}
\alias{on_sample_percentage_spinbutton_changed}
\alias{on_save_as_activate}
\alias{on_save_button_clicked}
\alias{on_save_menu_activate}
\alias{on_score_radiobutton_toggled}
\alias{on_sensitivity_radiobutton_toggled}
\alias{on_summary_radiobutton_toggled}
\alias{on_svm_radiobutton_toggled}
\alias{on_tools_cluster_activate}
\alias{on_tools_data_activate}
\alias{on_tools_evaluate_activate}
\alias{on_tools_explore_activate}
\alias{on_tools_log_activate}
\alias{on_tools_model_activate}
\alias{on_tools_sample_activate}
\alias{on_tools_variables_activate}
\alias{on_tooltips_activate}
\alias{on_twoclass_radiobutton_toggled}
\alias{on_unsupervised_radiobutton_toggled}
\alias{on_variables_toggle_ignore_button_clicked}
\alias{on_variables_toggle_input_button_clicked}
\alias{open_odbc_set_combo}
\alias{quit_rattle}
\alias{update_comboboxentry_with_dataframes}
\title{Interal Rattle user interface callbacks.}
\description{These are exported from the package so that the GUI can
  access the callbacks.}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
\keyword{internal}
