\name{rattle.print.rpart}
\alias{rattle.print.rpart}
\title{
  Modified printing of rpart decision tree
}
\description{
  
  Display a textual veiw of an rpart decision tree.
  
}
\usage{
rattle.print.rpart(x, minlength = 0, spaces = 2, cp,
                   digits = getOption("digits"), ...)
}
\arguments{
  
  \item{x}{}

  \item{minlength}{}

  \item{spaces}{}

  \item{cp}{}

  \item{digits}{}

  \item{...}{}

}
\details{

  Print decision tree.

}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
\keyword{tree}
